Fixed problem with value truncation

+++ ui/ui_main.js
@@ -100,9 +100,9 @@ angular.module('main', ['ngRoute'])
       var value;
       if($scope.value.indexOf('$') !== -1) {
         var no_dollar = $scope.value.slice(1);
-        value = parseInt(no_dollar) * 100;
+        value = parseFloat(no_dollar) * 100;
       } else {
-        value = parseInt($scope.value) * 100;
+        value = parseFloat($scope.value) * 100;
       }
 
       var new_expense = {
